worker/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
openai==1.56.1
anthropic==0.40.0
pinecone-client==5.0.1
boto3==1.34.158
pyodbc==5.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
PyJWT==2.9.0
python-dotenv==1.0.1